♻️ Refactor: 예약 현황 쿼리 파일 queries 폴더로 이동 및 QUERY_KEYS 추가 후 적용
