+++ src/shared/constants/queryKey.ts
@@ -14,4 +14,16 @@ export const QUERY_KEYS = {
   /** 내 예약 내역 조회 */
   MY_RESERVATIONS: (status?: string, size?: number) =>
     ['my-reservations', status ?? 'all', size ?? 'default'] as const,
+  /** 내 체험 목록 조회 */
+  MY_ACTIVITIES: (cursorId?: number, size?: number) =>
+    ['my-activities', cursorId ?? null, size ?? 10] as const,
+  /** 내 체험 월별 예약 현황 조회 */
+  RESERVATION_DASHBOARD: (activityId: number | null, year: string, month: string) =>
+    ['my-activities', activityId, 'reservation-dashboard', year, month] as const,
+  /** 내 체험 날짜별 예약 스케줄 조회 */
+  RESERVED_SCHEDULE: (activityId: number | null, date: string | null) =>
+    ['my-activities', activityId, 'reserved-schedule', date] as const,
+  /** 내 체험 예약 목록 조회 */
+  RESERVATIONS: (activityId: number | null, scheduleId: number, status: string) =>
+    ['my-activities', activityId, 'reservations', scheduleId, status] as const,
 } as const;
